% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{Counter.}
\usage{
counter(plan = "developer", day = NA, minute = NA)
}
\arguments{
\item{plan}{API usage plan.  Possible values are developer (500 calls a day 10
calls a minute, drizzle (5000 calls a day 100 calls a minute) shower (100000
calls a day 1000 a minute), or custom (see parameters day and minute).}

\item{day}{Custom daily API usage limit.}

\item{minute}{Custom minute API usage limit.}
}
\value{
Returns a counter object.
}
\description{
Counts API calls and tracks limits
}
\details{
Counter is a constructor function that returns a counter object for use by a
scheduler object.  The counter uses reference semantics so that multiple
schedulers may read and write to the counter object.
}
\examples{
counter(plan='drizzle')
}
