% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan}
\alias{plan}
\alias{plan.scheduler}
\title{Plan a schedule for executing a task}
\usage{
plan(scheduler, ...)

\method{plan}{scheduler}(scheduler, ...)
}
\arguments{
\item{scheduler}{A \code{\link{scheduler}} object.}

\item{...}{Arguments passed to \code{\link{seq.POSIXt}}.}
}
\value{
The schedule
}
\description{
Sets a schedule using strptime
}
\section{Methods (by class)}{
\itemize{
\item \code{scheduler}: convenience wrapper around seq.POSIXt.
}}

\examples{
plan(scheduler(), '1 hour') # sample every hour
plan(scheduler(), '30 min') # sample every 30 minutes
}
\seealso{
\code{\link{strptime}}, \code{\link{seq.POSIXt}}
}
