% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{scheduler}
\alias{scheduler}
\title{Schedules wunderscraper}
\usage{
scheduler(counter)
}
\arguments{
\item{counter}{A \code{\link{counter}} object}
}
\value{
Returns a scheduler object.
}
\description{
Uses a schedule and counter to schedule sampling and ensure wunderscraper
remains within API usage limits.
}
\details{
Scheduler is a constructor function that returns a scheduler object for use
in a wunderscrape process.

Scheduler has methods for managing the schedule: \code{\link{plan}}, and
\code{\link{sync}}.
}
\examples{
scheduler(counter(plan='drizzle'))
}
\seealso{
\code{\link{plan.scheduler}}, \code{\link{sync.scheduler}}
}
