% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apikey.R
\name{setApiKey}
\alias{setApiKey}
\title{Set API key.}
\usage{
setApiKey(key)
}
\arguments{
\item{key}{Valid Wunderground API key}
}
\value{
TRUE if setApiKey succeded
}
\description{
Sets the API key to the value of the argument.
}
\details{
Does not check if key is valid.
}
\examples{
setApiKey('1a2b3c4d5e6f7g8h9i0j')
}
