% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zctaRel}
\alias{zctaRel}
\title{Relationships for zip codes}
\format{A data frame with 44410 rows and 5 variables:
\describe{
  \item{ZCTA5}{zip code.}
  \item{COUNTY}{3 character string indicating county Federal Information
    Processing Standard (FIPS) code}
  \item{STATE}{2 character string indicating state FIPS code}
  \item{GEOID}{5 character string concatonating STATE and COUNTY}
  \item{COPOP}{county population}
}}
\source{
\url{https://www2.census.gov/geo/docs/maps-data/data/rel/zcta_county_rel_10.txt}
}
\usage{
zctaRel
}
\description{
A dataset containing population and land area of zip codes and administrative
boundaries
}
\examples{
summary(zctaRel)
}
\keyword{datasets}
