% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apikey.R
\name{setApiKey}
\alias{setApiKey}
\title{Set API key.}
\usage{
setApiKey(key, f = NULL)
}
\arguments{
\item{key}{A valid Wunderground API key}

\item{f}{A string indicating a text file containing only an API key}
}
\value{
TRUE if setApiKey succeded
}
\description{
Sets the API key to the value of the argument.
}
\details{
If \code{f} is not missing, then reads the API key from a file, else sets
key to the \code{key} parameter.  Does not check if key is valid.
}
\examples{
setApiKey('1a2b3c4d5e6f7g8h9i0j')
}
